Overview! And initial commit of the design docs.
This doesn't tell anyone anything specific; it's just a (very basic)
overview of what the engine is intended for. Nothing about how any part of
it works. Nothing on how anything fits together. As it says a *lot*, most
(nearly all) of it is going to be up to the people putting the game
together; this is just a (basic) framework and the overview document tells
that.

This is also the initial commit for the design_docs branch. As its name
says, the goal of the branch is to have a (usable) complete set of design
documents that will (hopefully) allow the engine to be designed to be as
useful as possible (by being bare-bones and minimal wherever it can be)
without having to deviate from them too much.

I'm not expecting to stick to them strictly; best laid plans and all that.
At worst though they'll detail how things *should have been*.

*NO CODE IN THIS BRANCH*; it's for documentation (design and otherwise)
only.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>