Added an in-game mockup! And state information.
Well, actually it's a screenshot from Sonic Outbreak but it'll serve my
purpose here. This is essentially how the base "game" is going to look. It
indicates on what layer(s) things are drawn on by default.

Also started on a document detailing a bit more about what the states are
supposed to do. They'll also list a bit more about how they work, as time
goes on and features and functionality are added (and bugs squashed!).

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>