Design of levels.
Not as in "how to design a good level", but "how level scenes should be
structured" - what nodes and scripts to use and how they inherit other
scripts. Not particularly complete at the moment - as things are coded it
will give better insight about what should be (or not be) in the document.

At some point I'm going to need to create an index/TOC that'll link to each
of these individual documents/images.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>