Collectibles!
Added a brief document about them and how they should work. This is pretty
vague at the mo; while some of this will be added to as experience and
knowledge of actually coding things grows, most of it will up to any game
makers.

Also renamed the design docs to hopefully follow some kind of structure.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>