The final commit...
...before merge!

It's not finished - won't ever be "finished" - but it is *ready* to use and
abuse (and to change whenever new things crop up, or finding problems and
so on). Hopefully it's ready enough to build a very basic skeleton from. If
not, not really a problem - just learn from things and add to this.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>