First attempt at player movement...
Pretty buggy and not doing much, but there's some code in there - it just
needs working and refining until it does (more or less) what it's supposed
to be doing.

Remember, movement state is the *direction to move in*, which is different
to the *direction currently moving in*. Current code isn't doing that (it
doesn't do much of anything right now though!).

The HUD uses the SVG assets directly now. Shouldn't make much difference to
anything otherwise though.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>