Working? Well, one step closer...
The print statements appear to look the way they should if nothing else
and animation seems to play appropriately. Does it mean it works? Not
quite, that's for sure, but it's looking more promising. In any case, all
this code is going to take quite a battering when more things are
implemented, especially floor/jumping checks.

All the limits - speed, jumping etc. - are going to be defined as export
variables so can be changed in the character scene directly without having
to edit its script.

There's also a test "platform" (just a StaticBody2D) for use with testing.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>