Cutscene controls and accelerate/decelerate helpers.
There's now a debugging control to toggle being in the cutscene state or
not.

Also, functions to work out multipliers for acceleration and deceleration
have been made. Right now, they're nearly empty, just default stuff. I can
easily see code for acceleration and deceleration getting quite complex,
dealing with multiple states and playable characters, so having it in just
one place will (hopefully!) make it a bit easier to deal with. There's also
a function to get the maximum speed multiplier and one to enforce limits on
the character's speed.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>