Jumping - it's a start, isn't it?
*Very* basic jumping model. Not fast, very floaty, but it's a start! The
real hard work is going to make it "easily" adjustable to taste by
developers, and to fit in with the default engine physics.

The current jump formula works like this: velocity.y has to be a value that
is larger than the gravity value. The difference is the (number of pixels
up+number of pixels down) in the jump.

Also, minor workarounds for a Godot random number bug, and added more info
to the debug HUD. Altered some default animations a bit.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>