Import file changes, tilesets and other things!
Been getting crashes with the last few betas of Godot editing the tileset
for the demo level, need to check that out more and see if it's been
reported yet (probably).

Got more checks for cutscene state in there (in some circumstances it was
still possible to move after cutscene state had been enabled).

The .import file format has had some changes again with images/textures it
seems, so all the relevant files have changed.

And added a starting point to the (current) demo level.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>