Starting to look into loops.
And... well, yeah. Dealing with getting the player around them is going to
be tricky. Wondering how Marmito did it with his little test level. Going
to try using some kind of automated "trick" depending on how fast the
player moves; move fast enough past a certain point and the player will be
automatically taken around the loop. May need to adjust the max angle (or
have it be automatically adjusted when needed).

Collision shapes are likely to not be my friend during all this...

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>