Loopy loops?
Eventually so. Going to be making some pretty major changes to the player
physics at some point soon I think. But as it is, (very) initial loop
implementation has been started. Going to use Marmito's loop and its code
as a base point. Which is, for this, a loop with several collision shapes
(floor, top half of loop, bottom left quarter, bottom right quarter) with
three Area2Ds that trigger when the player passes them by (and sets up the
collision shapes as required).

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>