Dead...?
As it stands, maybe. Trying to actually learn about KinematicBody2D's
default functionality and trying to use just that to implement something
like Sonic-ish physics. NOT integrated here yet, as it barely works any
better (at the moment at least).

The player code (assuming what I'm trying works, and I'm not sure about
the possibility of that just yet) is going to be removed and replaced. But
at least some of the non-player code is still usable/functional, so shall
be kept.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>