Simplest solution? A part of it at least. Small, incremental changes that are tested is the way to go here. Keep going until it's ready to be merged into master. Definitely do NOT just rewrite the thing from scratch.
* Hooked up the movement direction to input strength.
* Changed velocity to use it.
* Removed one of the sprite facing lines, and changed the other to take
  velocity's X direction into account. Also moved it into the general
  physics processing function to allow grinding, being in the air etc. to
  be taken into account.

It seems to work at least as much as the older code did (not that very much
has been changed yet). No issues fixed as of yet, though. Not that this is
intended to, just provide a platform for doing so.

Not for this branch, but future plans include moving all the player control
code into _input.

An "interesting" side-effect is that the walking down walls bug has a new
wrinkle. While walking down a wall the player is facing up. Well, that's
for when I open a branch to fix that issue.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>