It's all in the name.
* Renamed state flags to use is_ in their names to better signify what it
  is they're meant to do. Renamed some related variables to read better
  about what they do, too.
* Renamed some functions to be more readable.
* Made it so you can reset the character without having to reset the game.
* Potential bug-fix with handling boosting.

The goal here is to make things more friendly/readable when it comes to
readjusting input controls. The bug-fix may make it easier to implement
what I want, possibly without having to have extra variables (as is in the
current Boost_fixing.txt), we shall see. It's rarely ever that simple!

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>