Input changes (#19)
* Changes to boost handling, capability.

At least, that's the plan!

Wondering about how to get away from the use of Input in player_sonic.gd,
but moving all input-related functionality into player_generic.gd and have
the player-specific scripts handle how that input is used. Leaning towards
a counter of how many frames the boost key has been held down? Dunno, yet.
As few functionality changes as possible, anything like that will be dealt
with down the line.

Anyway:
* Added capability flags for the player being able to boost, fly or glide.
  They're simple exported booleans.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>

* It's all in the name.

* Renamed state flags to use is_ in their names to better signify what it
  is they're meant to do. Renamed some related variables to read better
  about what they do, too.
* Renamed some functions to be more readable.
* Made it so you can reset the character without having to reset the game.
* Potential bug-fix with handling boosting.

The goal here is to make things more friendly/readable when it comes to
readjusting input controls. The bug-fix may make it easier to implement
what I want, possibly without having to have extra variables (as is in the
current Boost_fixing.txt), we shall see. It's rarely ever that simple!

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>

* Better the Level you node.

* In preparation for genericising the levels and their code, renamed their
  root nodes to Level (still type Node2D). Adjusted scripts as needed.

Lots to be done there, especially as the UI code is intermixed with so much
other code.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>

* Boosting input/processing separated.

* Made is_boosting into a counter, moved the input code for boost into
  player_generic.gd, the processing code remains in the player-specific
  script.
* Jumping while boosting is no longer permitted. It's still possible to
  boost while jumping, however.

Still to do - crouch, spindash, jump etc. It's a start! It doesn't work
perfectly, but it'll do for now.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>

* Separate stomping.

* Separate the stomp input code from the processing code.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>

* Jump!

Separated out the jumping logic.

* Added an is_jumping flag, which handles processing of jump logic so long
  as the jump button is held down.
* Moved the input code for jumping to _input in player_generic.gd.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>

* Crouching.

* Moved setting is_crouching to _input, player_generic.gd.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>

* Tidying up around the edges.

* Moved the reset/restart functions/functionality into player_generic.gd
* Changed the last Input uses in player_<name>.gd to use flags instead.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>