Initial observations?
The player character collision shape is a CapsuleShape2D, with height and
radius properties. Documenting a little bit about magic numbers and their
relationship (if any) to the collision shape size. Not entirely confident
about that, so will be checking them out.

height is 20, radius is 10.

I'm a *little* confident about the "21" magic number as it (as far as I can
see) seems only to be used with y-component calculations.

A downside of this is that collision shapes can only be CapsuleShape2D for
the player characters.

* Separated player sprites from player scene.
* Initial "magic number" thoughts.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>