Investigate magic numbers (#24)
* Initial observations?

The player character collision shape is a CapsuleShape2D, with height and
radius properties. Documenting a little bit about magic numbers and their
relationship (if any) to the collision shape size. Not entirely confident
about that, so will be checking them out.

height is 20, radius is 10.

I'm a *little* confident about the "21" magic number as it (as far as I can
see) seems only to be used with y-component calculations.

A downside of this is that collision shapes can only be CapsuleShape2D for
the player characters.

* Separated player sprites from player scene.
* Initial "magic number" thoughts.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>

* Magic number experimentation.

Reasonably confident the magic number 21 has been eliminated in a couple
of places, not so sure about the others, but from what I can tell (so far)
they may be accurate. Time will tell; the biggest test will be adding more
playable characters.

21 was height-related (as far as I can tell). Not sure yet where radius
will come into it, shall investigate further.

* Added variables to hold the collider shape height and radius.
* Eliminated (as far as I can tell) the 21 magic number.
* Separated out the boost sprites from the player scene.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>

* The mystery of "14".

14 seems to be being used somewhere to stop sideways collisions. It seems
to be the collision radius * 1.4. Not sure how accurate that is, but going
to put it in here and see how it works when it comes to other characters.

* Eliminated the 14 magic number (possibly).

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>

* The mysteries of magic numbers.

* Changed one calculation to another which may be more accurate. Only time
  will tell.

Not sure how much of this is going to work with other characters, but we
will have to see.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>